logic spread among different files and classes, minor fixes
