+++ daas/daas_app/utils/charts/__init__.py
@@ -1,3 +1,4 @@
 from .bar_charts import samples_per_size, samples_per_elapsed_time
 from .pie_charts import samples_per_type
 from .bar_zoom_chart import bar_datazoom_slider
+from .statistics_manager import StatisticsManager
