+++ daas/daas_app/utils/charts/bar_zoom_chart.py
@@ -1,19 +1,12 @@
 from pyecharts.charts import Line
 from pyecharts import options as opts
-import bottleneck
-from typing import List
 
-from ..statistics_manager import StatisticsManager
+from .statistics_manager import StatisticsManager
 from ..configuration_manager import ConfigurationManager
 
 
-def simple_moving_average(series: List[int], window: int = 7):
-    """ Calculates the simple moving average. The <window> parameter indicates how much values to consider. """
-    return [round(move_mean_value, 1) for move_mean_value in bottleneck.move_mean(series, window=window, min_count=1)]
-
-
 def bar_datazoom_slider() -> Line:
-    groups = [StatisticsManager().get_sample_count_per_upload_date(file_type) for file_type in
+    groups = [StatisticsManager().get_sample_counts_per_upload_date(file_type) for file_type in
               ConfigurationManager().get_identifiers()]
     chart = Line()
     chart.add_xaxis(groups[0].captions)
@@ -22,11 +15,10 @@ def bar_datazoom_slider() -> Line:
                         group.counts,
                         linestyle_opts=opts.LineStyleOpts(width=2),
                         label_opts=opts.LabelOpts(is_show=False))
-        chart.add_yaxis(f'{group.file_type} SMA (7 days)',
-                        simple_moving_average(group.counts),
+        chart.add_yaxis(group.simple_moving_average_legend,
+                        group.simple_moving_average,
                         linestyle_opts=opts.LineStyleOpts(width=2, type_='dashed'),
                         label_opts=opts.LabelOpts(color="black", background_color="white"))
     chart.set_global_opts(title_opts=opts.TitleOpts(title="Samples per upload date"),
                           datazoom_opts=opts.DataZoomOpts())
-    #.set_series_opts(label_opts=opts.LabelOpts(is_show=False))
     return chart
